requirements.txt
python-telegram-bot>=20.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
matplotlib>=3.5.0
SQLAlchemy==2.0.23
alembic==1.12.1
flask==2.3.3
werkzeug==2.3.7
openai>=1.0.0
pydub>=0.25.1 

aiogram
google-generativeai
pytz